fix: OOM Kill alert considers container restart
The `kube_pod_container_status_last_terminated_reason`
gives us the reasons of the last terminated container.

If the container restarts due to OOM Kill but then runs succesfully,
the metric will continue to report the OOM Kill for the same pod.

Therefore, we need to also consider whether a container restart
happened in the last 5 minutes.