refactor(helm): rename and update priority class configuration
Remove unused helper templates and rename the priority class
configuration from `priorityClass` to `prirotiyClassNodeCritical` for
clarity. Update the priority class value to a higher number to reflect
its critical importance. Remove the old `priorityClass.yaml` and add
`prirotiyClassNodeCritical.yaml` to align with the new naming
convention. This change improves the organization and clarity of the
Helm chart configuration.